Extract data loader and event handlers to dedicated files
