+++ src/trackers/arcdps_tracker.cpp
@@ -1,28 +1,27 @@
 #include "arcdps_tracker.h"
-#include "../core/log_proofs.h"
+#include "../core/player_manager.h"
+
+extern bool unofficialExtrasEnabled;
 
 bool ArcdpsTracker::IsAvailable() const {
-	return !LogProofs::unofficialExtrasEnabled;
+	return !unofficialExtrasEnabled;
 }
 
-void ArcdpsTracker::Initialize() {
-}
-
-void ArcdpsTracker::Shutdown() {
-}
+void ArcdpsTracker::Initialize() {}
+void ArcdpsTracker::Shutdown() {}
 
 void ArcdpsTracker::OnPlayerJoin(const PlayerInfo& player) {
-	LogProofs::AddPlayerFromTracker(player);
+	PlayerManager::AddPlayer(player);
 }
 
 void ArcdpsTracker::OnPlayerLeave(const PlayerInfo& player) {
-	LogProofs::RemovePlayerFromTracker(player);
+	PlayerManager::RemovePlayer(player);
 }
 
 void ArcdpsTracker::OnSelfDetected(const PlayerInfo& self) {
-	LogProofs::SetSelfFromTracker(self);
+	PlayerManager::SetSelf(self);
 }
 
 void ArcdpsTracker::OnSquadClear() {
-	LogProofs::ClearPlayers();
+	PlayerManager::ClearPlayers();
 }
